# Improve performance of computing columns of matrices over GF(2)