@GiacomoPope 2024-06-05
Thanks for the tips -- I was lazy and copy pasted old columns code, so these changes could be applied elsewhere too but that's for another day!